+++ aws-lambda/lib/extended-go-function.ts
@@ -29,11 +29,7 @@ export class ExtendedGoFunction extends GoFunction {
   readonly deployment?: FunctionDeployment;
 
   constructor(scope: Construct, id: string, props: ExtendedGoFunctionProps) {
-    const logGroup = configureLogGroupForFunction(
-      scope,
-      `${id}LogGroup`,
-      props,
-    );
+    const logGroup = configureLogGroupForFunction(scope, id, props);
 
     super(scope, id, {
       logGroup,